ref(feedback): extract screenshot fetching into a hook
